+++ CHANGELOG.md
@@ -155,6 +155,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+- Fixed allow `ex.ColliderComponent` to not have a collider
 - Fixed issue where collision events were not being forwarded from individual colliders in a `ex.CompositeCollider`
 - Fixed issue where `ex.CompositeCollider`'s individual colliders were erroneously generating pairs
 - Fixed issue where `GraphicsOptions` `width/height` could not be used to define a `Sprite` with equivalent `sourceView` and `destSize` ([#1863](https://github.com/excaliburjs/Excalibur/issues/1863))